RAG/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0

